4.22*: blogilistan laajennus, step11

+++ controllers/blogs.js
@@ -38,7 +38,7 @@ blogsRouter.post('/', async (request, response) => {
 
     const decodedToken = jwt.verify(request.token, process.env.SECRET)
     if (!request.token || !decodedToken.id) {
-        return response.status(401).json({ error: 'token missing or invalid' })
+        return response.status(401).json({ error: 'token missing or invalid' }).end()
     }
 
     const user = await User.findById(decodedToken.id)
@@ -87,9 +87,26 @@ blogsRouter.put('/:id', async (request, response) => {
 
 // Delete one blog based on id
 blogsRouter.delete('/:id', async (request, response) => {
-    await Blog.findByIdAndRemove(request.params.id)
-    // Status 204 include both blog remove and requested blog not exist
-    response.status(204).end()
+    const body = request.body
+    const decodedToken = jwt.verify(request.token, process.env.SECRET)
+
+    if (!request.token || !decodedToken.id) {
+        return response.status(401).json({ error: 'token missing or invalid' }).end()
+    }
+
+    console.log('request.params.id "request blog"', request.params.id)
+    const blog = await Blog.findById(request.params.id)
+    console.log('blog.user "created by"', blog.user)
+    console.log('decodedToken.id "delete by"', decodedToken.id)
+
+    // Only blog creator has right to delete blog
+    if (blog.user.toString() === decodedToken.id.toString()) {
+        await Blog.findByIdAndRemove(request.params.id)
+        // Status 204 include both blog remove and requested blog not exist
+        response.status(204).end()
+    } else {
+        return response.status(401).json({ error: 'Unauthorized user' }).end()
+    }
 })
